+++ src/base/styles/colors.ts
@@ -1,0 +1,4 @@
+export const COLORS = {
+  GRAY: '#848484',
+  WHITE: '#ffffff',
+};