+++ src/config/linking.ts
@@ -42,6 +42,9 @@ const routesConfig = {
         MY_CASHBACK_MY_WALLET: {
           path: 'wallet-cashback',
         },
+        PrimeLP: {
+          path: 'prime',
+        },
       },
     },
   },