+++ src/pages/ProductDetail/components/ProductAddToCart/ProductAddToCart.tsx
@@ -9,7 +9,7 @@ import { useRemoteConfig } from '../../../../hooks/useRemoteConfig';
 import { useCart } from '../../../../context/CartContext';
 import EventProvider from '../../../../utils/EventProvider';
 import { useProductDetailStore } from '../../../../zustand/useProductDetail/useProductDetail';
-import { ModalBag } from '../../../../modules/ProductDetail/components/ModalBag';
+import { ModalBag } from '../../../../common/components/ModalBag/ModalBag';
 
 function ProductAddToCart() {
   const { getString } = useRemoteConfig();