+++ src/utils/testProps.ts
@@ -1,7 +1,12 @@
 import { Platform } from 'react-native';
+import { platformType } from './platformType';
 
-const testProps = (testID: string) => {
-  if (Platform.OS === 'ios') {
+const prefix = 'com.usereserva:id/' as const;
+
+const testProps = (value: string) => {
+  const testID = value.startsWith(prefix) ? value : `${prefix}${value}`;
+
+  if (Platform.OS === platformType.IOS) {
     return {
       testID,
       accessible: false,
@@ -9,6 +14,7 @@ const testProps = (testID: string) => {
   }
 
   return {
+    testID,
     accessible: true,
     accessibilityLabel: testID,
   };